Implement ARM and Thumb instruction interpretation and handling
- Added `arm_instruction.rs` and `thumb_instruction.rs` modules to handle ARM and Thumb instruction interpretation.
- Implemented basic structure for various ARM instructions including load/store, data processing, and branch instructions.
- Created `arm_interpret` and `arm_decode` functions for interpreting and decoding ARM instructions.
- Introduced `thumb_interpret` and `thumb_decode` functions for Thumb instruction handling.
- Removed the old CPU implementation in `cpu_old.rs` to streamline the codebase.
- Updated references in the emulator to accommodate the new instruction handling structure.
- Fixed minor issues in the cartridge and emulator modules related to register handling.

+++ core/lunaris_emu/src/emulator/read_arm9.rs
@@ -56,7 +56,7 @@ impl Emulator {
             0x040000EC => self.dma_fill[3],
             0x04000180 => self.ipc_sync_nds9.read().into(),
             0x040001A4 => self.cart.get_romctrl(),
-            0x04000208 => self.int9_reg.ime as u32,
+            0x04000208 => self.int9_reg.ime,
             0x04000210 => self.int9_reg.irq_enable,
             0x04000214 => self.int9_reg.irq_flags,
             0x04000240 => self.gpu.get_vramcnt_a() as u32,
